feat: add unit test and cleanup code
-add unit test to test user service
-do some cleanup on code
