--- UPS.Assessment/Infrastructure/Helpers/IServiceResponse.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace UPS.Assessment.Infrastructure.Helpers;
-
-public interface IServiceResponse
-{
-    IReadOnlyCollection<IResponseError> Errors { get; }
-    string? Message { get; }
-    bool Succeeded { get; }
-    int ItemsCount { get; }
-}
-
-public interface IServiceResponse<out T> : IServiceResponse
-{
-    T? Data { get; }
-}
